refactor integration test
